Inventory (Code)/inventory-master/requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
cycler==0.10.0
DateTime==4.3
Django==3.2.1
django-advanced-filters==1.3.0
django-allauth==0.41.0
django-braces==1.14.0
django-crispy-forms==1.11.2
django-totalsum-admin-django3==1.0.0
future==0.18.2
html5lib==1.1
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
packaging==20.9
Pillow==8.2.0
pip-review==1.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
randomcolor==0.4.4.5
reportlab==3.5.67
scikit-learn==0.24.2
scipy==1.6.3
simplejson==3.17.2
simpy==4.0.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.5
zope.interface==5.4.0
